pyproject.toml
[project]
name = "obsidian-link-extractor"
version = "0.1.0"
description = "Extract, summarize, and tag links from Obsidian daily notes"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "boto3>=1.34.0",
    "beautifulsoup4>=4.12.0",
    "pymupdf>=1.24.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
link-extractor = "link_extractor.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
